Fix types; move utilities to common
