+++ src/plugins/share/common/index.ts
@@ -6,4 +6,11 @@
  * Side Public License, v 1.
  */
 
-export { LocatorDefinition, LocatorPublic, useLocatorUrl } from './url_service';
+export {
+  LocatorDefinition,
+  LocatorPublic,
+  useLocatorUrl,
+  RedirectOptions,
+  formatSearchParams,
+  parseSearchParams,
+} from './url_service';
