+++ src/plugins/share/common/url_service/index.ts
@@ -8,3 +8,4 @@
 
 export * from './url_service';
 export * from './locators';
+export * from './redirect';
