+++ src/plugins/share/public/index.ts
@@ -9,7 +9,7 @@
 export { CSV_QUOTE_VALUES_SETTING, CSV_SEPARATOR_SETTING } from '../common/constants';
 
 export { LocatorDefinition, LocatorPublic, KibanaLocation } from '../common/url_service';
-export { parseSearchParams, formatSearchParams } from './url_service';
+export { parseSearchParams, formatSearchParams } from '../common/url_service';
 
 export { UrlGeneratorStateMapping } from './url_generators/url_generator_definition';
