+++ src/plugins/share/public/plugin.ts
@@ -18,8 +18,9 @@ import {
   UrlGeneratorsSetup,
   UrlGeneratorsStart,
 } from './url_generators/url_generator_service';
-import { UrlService } from '../common/url_service';
-import { formatSearchParams, RedirectManager, RedirectOptions } from './url_service';
+import { UrlService, formatSearchParams } from '../common/url_service';
+import type { RedirectOptions } from '../common';
+import { RedirectManager } from './url_service';
 
 export interface ShareSetupDependencies {
   securityOss?: SecurityOssPluginSetup;
